requirements.txt
flask
llama-cpp-python
requests
gunicorn
